update flake & upload.sh

+++ scripts/upload.sh
@@ -5,13 +5,24 @@ BASEURL="https://cdn.qeden.systems"
 DATE=$(date -u "+%Y%d%m")
 PKG="nixos-asahi-${DATE}.zip"
 RESULT=$(realpath "${BASEDIR}"/result)
-ROOTSIZE=$(cat ${BASEDIR}/result/.tag_rootsize)
+ROOTSIZE=$(cat "${RESULT}"/.tag_rootsize)
 VERSION_TAG=$(cat "${BASEDIR}"/.version_tag)
+
+confirm() {
+  while true; do
+    read -r -n 1 -p "${1:-Continue?} [y/n]: " REPLY
+    case $REPLY in
+      [yY]) echo ; return 0 ;;
+      [nN]) echo ; return 1 ;;
+      *) printf " \033[31m %s \n\033[0m" "invalid input"
+    esac
+  done
+}
 
 if [[ -e $RESULT/nixos-asahi-${DATE}.zip ]]; then
   PKG="nixos-asahi-${DATE}.zip"
-else
-  PKG=$(basename $(stat -f "%a %N" "$RESULT"/package/nixos-asahi-*.zip | sort -nr | head -n 1 | awk -F' ' '{print $2}'))
+# else
+  # PKG=$(basename $(stat -c "%a %N" "$RESULT"/package/nixos-asahi-*.zip | sort -nr | head -n 1 | awk -F' ' '{print $2}'))
 fi
 
 upload() {
@@ -19,10 +30,9 @@ upload() {
     cp -a "$RESULT"/package/"$PKG" /tmp/
     sudo chmod 644 /tmp/"$PKG"
   fi
-
-  rclone copy --progress /tmp/"$PKG" r2:nixos-asahi && echo "Success! $PKG uploaded to bucket."
-
-  rm -rf /tmp/"$PKG"
+  rclone copy --progress /tmp/"$PKG" r2:nixos-asahi || exit 1
+  [[ $? == 0 ]] && echo "Success! $PKG uploaded to bucket."
+  [[ -e /tmp/"$PKG" ]] && rm -rf /tmp/"$PKG"
 }
 
 update_installer_data() {
@@ -35,16 +45,15 @@ increment_version() {
 }
 
 main() {
+  confirm "Begin upload?" || exit 0
   upload
-  printf "\nUpdate package version and push to git? (press enter x2)"
-  read -r
-  read -r
+  [[ $? == 0 ]] && confirm "Update package version and push to git?" || exit 0
   increment_version
   update_installer_data
   git add "${BASEDIR}"/data/installer_data.json "${BASEDIR}"/.version_tag
   git commit -m "update package: ${VERSION}"
   git tag "${VERSION}"
-  git push
+  git push -u origin "${VERSION}"
 }
 
 main && exit
